refactor(a11y): split axe gate test to comply with max-lines-per-function
